requirements.txt
python-dotenv==1.0.1
keyring==25.2.1
aiohttp==3.9.5
colorama==0.4.6
customtkinter==5.2.2
APScheduler==3.10.4
windows-toasts==1.3.1